+++ apps/diffusion/package.json
@@ -21,7 +21,6 @@
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "react-image-gallery": "0.8.12",
-    "react-slick": "^0.23.2",
     "react-viewer": "^2.8.0",
     "reactstrap": "^7.1.0",
     "shave": "^2.5.2",
